fix(workbench): per-version tags in compare + redesign Add Tags modal [TH-6228]
