+++ frontend/src/components/ModalWrapper/ModalWrapper.jsx
@@ -32,6 +32,7 @@ export default function ModalWrapper({
   actionBtnProps = {},
   cancelBtnProps = {},
   onCancelBtn,
+  paperSx = {},
 }) {
   const theme = useTheme();
   return (
@@ -50,6 +51,7 @@ export default function ModalWrapper({
           display: "flex",
           flexDirection: "column",
           gap: theme.spacing(2),
+          ...paperSx,
         },
       }}
     >
@@ -163,4 +165,5 @@ ModalWrapper.propTypes = {
   modalWidth: PropTypes.string,
   actionBtnProps: PropTypes.object,
   cancelBtnProps: PropTypes.object,
+  paperSx: PropTypes.object,
 };
