+++ frontend/src/sections/workbench/createPrompt/VersionHistory/LabelDropdown/LabelSelectContent.jsx
@@ -1,22 +1,24 @@
 import {
   Box,
+  Button,
   Chip,
   Divider,
-  Popper,
   Typography,
   useTheme,
 } from "@mui/material";
 import PropTypes from "prop-types";
-import React, { useCallback, useMemo, useRef, useState } from "react";
+import React, { useCallback, useMemo, useState } from "react";
 import Iconify from "src/components/iconify";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import axios, { endpoints } from "src/utils/axios";
 import { enqueueSnackbar } from "notistack";
 import { PERMISSIONS, RolePermission } from "src/utils/rolePermissionMapping";
 import { useAuthContext } from "src/auth/hooks";
-import { getColorMap } from "../common";
-import LabelPopperContent from "./LabelPopperContent";
+import SvgColor from "src/components/svg-color";
 import { LoadingButton } from "@mui/lab";
+import { getColorMap } from "../common";
+import LabelList from "./LabelList";
+import CreateTagInput from "./CreateTagInput";
 
 const LabelSelectContent = ({
   promptId,
@@ -30,8 +32,6 @@ const LabelSelectContent = ({
   isFetchingNextPage,
   fetchNextPage,
 }) => {
-  const anchorRef = useRef(null);
-  const [open, setOpen] = useState(false);
   const [showCreateForm, setShowCreateForm] = useState(false);
   const [newLabelName, setNewLabelName] = useState("");
   const theme = useTheme();
@@ -68,7 +68,6 @@ const LabelSelectContent = ({
         queryKey: ["prompt-latest-version", promptId],
       });
       enqueueSnackbar("Labels updated successfully!", { variant: "success" });
-      setOpen(false);
       setSelectedLabelsList([]);
       onClose?.();
       onSuccess?.();
@@ -91,7 +90,7 @@ const LabelSelectContent = ({
         type: "all",
       });
       enqueueSnackbar("Label created successfully!", { variant: "success" });
-      setShowCreateForm(false);
+      // Keep the create input open for back-to-back adds.
       setNewLabelName("");
     },
   });
@@ -125,6 +124,11 @@ const LabelSelectContent = ({
     createLabel({ name: newLabelName.trim() });
   }, [newLabelName, createLabel]);
 
+  const handleCancelCreate = useCallback(() => {
+    setShowCreateForm(false);
+    setNewLabelName("");
+  }, []);
+
   const isLabelSelected = useCallback(
     (labelId) => selectedLabelsList.some((l) => l.id === labelId),
     [selectedLabelsList],
@@ -132,147 +136,136 @@ const LabelSelectContent = ({
 
   return (
     <>
-      <Divider orientation="horizontal" sx={{ marginX: "-12px", mb: 2 }} />
+      <Divider orientation="horizontal" sx={{ marginX: "-12px", mb: 1.5 }} />
 
-      <Box sx={{ position: "relative" }}>
-        <Box
-          ref={anchorRef}
-          onClick={() => setOpen((prev) => !prev)}
+      <Box
+        sx={{
+          position: "relative",
+          display: "flex",
+          alignItems: "center",
+          flexWrap: "wrap",
+          gap: 0.5,
+          minHeight: "40px",
+          padding: "5px 12px",
+          border: "1px solid",
+          borderColor: "divider",
+          borderRadius: "6px",
+        }}
+      >
+        <Typography
+          typography={"s3"}
+          fontWeight={"fontWeightMedium"}
           sx={{
-            position: "relative",
-            display: "flex",
-            alignItems: "center",
-            flexWrap: "wrap",
-            gap: 0.5,
-            minHeight: "40px",
-            padding: "8px 40px 8px 14px",
-            border: "1px solid",
-            borderColor: "divider",
-            borderRadius: "4px",
-            transition: "border-color 0.2s",
-            "&:hover": {
-              borderColor: "divider",
-            },
+            position: "absolute",
+            top: selectedLabelsList.length > 0 ? -8 : 11,
+            left: 14,
+            backgroundColor: "background.paper",
+            padding: "0 4px",
+            color: "text.secondary",
+            fontSize: selectedLabelsList.length > 0 ? "12px" : "14px",
+            transition: "all 0.2s",
+            pointerEvents: "none",
           }}
         >
-          <Typography
-            typography={"s3"}
-            fontWeight={"fontWeightMedium"}
-            color={"text.disabled"}
-            sx={{
-              position: "absolute",
-              top: selectedLabelsList.length > 0 ? -8 : 11,
-              left: 14,
-              backgroundColor: "background.paper",
-              padding: "0 4px",
-              color: "text.secondary",
-              fontSize: selectedLabelsList.length > 0 ? "12px" : "14px",
-              transition: "all 0.2s",
-              pointerEvents: "none",
-            }}
-          >
-            Select Tags
-          </Typography>
+          Selected Tags
+        </Typography>
 
-          {selectedLabelsList.map((label) => {
-            const colorMap = getColorMap(label?.name, theme);
-            return (
-              <Chip
-                key={label.id}
-                label={label.name}
-                size="small"
-                disabled={!hasDeployPermission}
-                onDelete={
-                  hasDeployPermission ? () => handleSelect(label) : undefined
-                }
-                deleteIcon={
-                  <Iconify
-                    icon="mdi:close"
-                    sx={{
-                      width: 14,
-                      height: 14,
-                    }}
-                  />
-                }
-                sx={{
-                  backgroundColor: colorMap?.backgroundColor,
+        {selectedLabelsList.map((label) => {
+          const colorMap = getColorMap(label?.name, theme);
+          return (
+            <Chip
+              key={label.id}
+              label={label.name}
+              size="small"
+              disabled={!hasDeployPermission}
+              onDelete={
+                hasDeployPermission ? () => handleSelect(label) : undefined
+              }
+              deleteIcon={
+                <Iconify
+                  icon="mdi:close"
+                  sx={{
+                    width: 14,
+                    height: 14,
+                  }}
+                />
+              }
+              sx={{
+                backgroundColor: colorMap?.backgroundColor,
+                color: colorMap?.color,
+                borderRadius: "100px",
+                "&:hover": {
+                  backgroundColor: colorMap?.hoverBackgroundColor,
+                },
+                "& .MuiChip-deleteIcon": {
                   color: colorMap?.color,
-                  borderRadius: "100px",
-                  zIndex: 10,
-                  "&:hover": {
-                    backgroundColor: colorMap?.hoverBackgroundColor,
-                  },
-                  "& .MuiChip-deleteIcon": {
-                    color: colorMap?.color,
-                  },
-                  typography: "s2_1",
-                  fontWeight: "fontWeightMedium",
-                }}
-              />
-            );
-          })}
-
-          <Box sx={{ flex: 1, minWidth: "50px" }} />
-
-          <Iconify
-            icon="eva:arrow-ios-upward-fill"
-            sx={{
-              position: "absolute",
-              right: 12,
-              top: "50%",
-              transform: `translateY(-50%) rotateX(${open ? 0 : -180}deg)`,
-              cursor: hasDeployPermission ? "pointer" : "default",
-              color: hasDeployPermission ? "text.primary" : "text.disabled",
-              width: 20,
-              height: 20,
-              transition: "transform 0.3s",
-            }}
-            onClick={(e) => {
-              if (!hasDeployPermission) return;
-              e.stopPropagation();
-              setOpen((prev) => !prev);
-            }}
-          />
-        </Box>
-
-        <Popper
-          open={open}
-          anchorEl={anchorRef.current}
-          placement="bottom-start"
-          style={{
-            width: anchorRef.current?.offsetWidth || 467,
-            zIndex: 1300,
-          }}
-        >
-          <LabelPopperContent
-            labels={labels}
-            isPending={isPending}
-            isFetchingNextPage={isFetchingNextPage}
-            isLabelSelected={isLabelSelected}
-            handleSelect={handleSelect}
-            showCreateForm={showCreateForm}
-            setShowCreateForm={setShowCreateForm}
-            newLabelName={newLabelName}
-            setNewLabelName={setNewLabelName}
-            handleCreateLabel={handleCreateLabel}
-            isCreatingLabel={isCreatingLabel}
-            onClose={() => setOpen(false)}
-            version={version}
-            fetchNextPage={fetchNextPage}
-          />
-        </Popper>
+                },
+                typography: "s2_1",
+                fontWeight: "fontWeightMedium",
+              }}
+            />
+          );
+        })}
       </Box>
 
-      <LoadingButton
-        disabled={!hasChanges}
-        loading={isAssigning}
-        color="primary"
-        variant="contained"
-        onClick={handleSave}
-        fullWidth
+      <LabelList
+        labels={labels}
+        isPending={isPending}
+        isFetchingNextPage={isFetchingNextPage}
+        isLabelSelected={isLabelSelected}
+        handleSelect={handleSelect}
+        version={version}
+        fetchNextPage={fetchNextPage}
+      />
+
+      {showCreateForm && (
+        <CreateTagInput
+          newLabelName={newLabelName}
+          setNewLabelName={setNewLabelName}
+          handleCreateLabel={handleCreateLabel}
+          isCreatingLabel={isCreatingLabel}
+          onCancel={handleCancelCreate}
+        />
+      )}
+
+      <Box
+        sx={{
+          display: "flex",
+          alignItems: "center",
+          justifyContent: "space-between",
+          gap: 1,
+          mt: 1.25,
+        }}
       >
-        Save
-      </LoadingButton>
+        <Button
+          size="small"
+          variant="outlined"
+          color="inherit"
+          disabled={showCreateForm || !hasDeployPermission}
+          onClick={() => setShowCreateForm(true)}
+          startIcon={
+            <SvgColor
+              src="/assets/icons/ic_add.svg"
+              sx={{ width: 16, height: 16 }}
+            />
+          }
+        >
+          <Typography typography="s3" fontWeight={500}>
+            Add custom tag
+          </Typography>
+        </Button>
+
+        <LoadingButton
+          size="small"
+          disabled={!hasChanges}
+          loading={isAssigning}
+          color="primary"
+          variant="contained"
+          onClick={handleSave}
+        >
+          Save
+        </LoadingButton>
+      </Box>
     </>
   );
 };
